requirements.txt
pandas
requests
beautifulsoup4
selenium
streamlit
scikit-learn
xgboost
tensorflow
keras
plotly
google-generativeai